+++ internal/gossiper/manual.go
@@ -70,7 +70,7 @@ func (g *Manual) Force(ctx context.Context) error {
 	if len(txs) == 0 {
 		return nil
 	}
-	b, err := chain.MarshalSignedTxs(txs)
+	b, err := chain.MarshalTxs(txs)
 	if err != nil {
 		return err
 	}
@@ -87,7 +87,7 @@ func (g *Manual) Force(ctx context.Context) error {
 
 func (g *Manual) HandleAppGossip(ctx context.Context, nodeID ids.NodeID, msg []byte) error {
 	actionRegistry, authRegistry := g.vm.ActionRegistry(), g.vm.AuthRegistry()
-	_, txs, err := chain.UnmarshalSignedTxs(msg, initialCapacity, actionRegistry, authRegistry)
+	_, txs, err := chain.UnmarshalTxs(msg, initialCapacity, actionRegistry, authRegistry)
 	if err != nil {
 		g.vm.Logger().Warn(
 			"AppGossip provided invalid txs",
